-Installer: cleanup
git-svn-id: http://svn.miranda-ng.org/main/trunk@5932 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ tools/installer_ng/cleanup.bat
@@ -1,7 +1,7 @@
 @echo off
 rem Cleanup
-rd /S /Q tmp
-rd /S /Q util
-rd /S /Q InnoNG_32
-rd /S /Q InnoNG_64
+if exist tmp rd /S /Q tmp
+if exist util rd /S /Q util
+if exist InnoNG_32 rd /S /Q InnoNG_32
+if exist InnoNG_64 rd /S /Q InnoNG_64
 rem end
